// src/modules/notifications/domain/enums/notification-type.enum.ts
export enum NotificationType {
  // Community
  POST_LIKE = 'POST_LIKE',
  POST_COMMENT = 'POST_COMMENT',
  COMMENT_LIKE = 'COMMENT_LIKE',
  COMMENT_REPLY = 'COMMENT_REPLY',
  POST_MENTION = 'POST_MENTION',
  
  // Reports (Admin)
  NEW_REPORT = 'NEW_REPORT',
  REPORT_RESOLVED = 'REPORT_RESOLVED',
  
  // News
  NEWS_PUBLISHED = 'NEWS_PUBLISHED',
  NEWS_FEATURED = 'NEWS_FEATURED',
  
  // Account
  WELCOME = 'WELCOME',
  PASSWORD_CHANGED = 'PASSWORD_CHANGED',
  
  // System
  SYSTEM_ANNOUNCEMENT = 'SYSTEM_ANNOUNCEMENT',
}